# integration-enabler-spring-v1/src/test/resources/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Static Swagger Info for Sample micro-service showing how to enable a Spring Boot v1.x application",
        "version": "1.0.1",
        "title": "Service Integration Enabler V1 Sample App (spring boot 1.x)"
    },
    "basePath": "/samplev1client",
...
    "paths": {
        "/instance/configured-port": {
            "get": {
                "tags": [
                    "Instance"
                ],
                "summary": "What port is this controller configured for",
                "description": "What port is this controller configured for",
                "operationId": "getPortUsingGET",
                "produces": [
                    "text/plain"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "URI not found"
                    }
                }
            }
        },
        "/instance/gateway-url": {
            "get": {
                "tags": [
                    "Instance"
                ],
                "summary": "What is the URI of the Gateway",
                "description": "What is the URI of the Gateway",
                "operationId": "getGatewayLocationUsingGET",
                "produces": [
                    "text/plain"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "URI not found"
                    },
                    "500": {
                        "description": "Internal Error Occurred"
                    }
                }
            }
        },
...
    },
...
}